wip2

+++ src/args.rs
@@ -68,7 +68,7 @@ pub struct Args {
     /// Given in chunks, size is chunks^2
     /// E.g. setting 4 will process 16 chunks at a time
     #[arg(long)]
-    pub batch_area_size: Option<u16>,
+    pub batch_area_size: Option<u32>,
 }
 
 fn validate_minecraft_world_path(path: &str) -> Result<String, String> {
